Refactor DiagnosticsRecorder
